refactor: 재고 등록 리팩토링
